fix(extension): move secret key files
