+++ apps/extension/src/app/ui/components/secret-key/secret-key-grid.tsx
@@ -8,11 +8,11 @@ export function SecretKeyGrid({ children }: SecretKeyGridProps) {
     <Grid
       gridTemplateColumns={{
         base: 'repeat(1, minmax(160px, 1fr))',
-        md: 'repeat(2, minmax(170px, 1fr))',
-        lg: 'repeat(3, minmax(170px, 1fr))',
+        sm: 'repeat(2, minmax(170px, 1fr))',
+        md: 'repeat(3, minmax(170px, 1fr))',
       }}
-      rowGap="space.03"
-      columnGap="space.03"
+      rowGap="space.02"
+      columnGap="space.02"
     >
       {children}
     </Grid>
