+++ apps/extension/src/app/ui/components/secret-key/secret-key-word.tsx
